rename parse_csv to parse_sync_file
