+++ src/neuroprocessing/parse_sync_file.py
